+++ selfdrive/controls/controlsd.py
@@ -160,6 +160,9 @@ class Controls:
     self.aReqValueMin = 0.
     self.aReqValueMax = 0.
 
+    self.left_lane_visible = False
+    self.right_lane_visible = False
+
     # TODO: no longer necessary, aside from process replay
     self.sm['liveParameters'].valid = True
 
@@ -341,7 +344,7 @@ class Controls:
     self.sm.update(0)
 
     all_valid = CS.canValid and self.sm.all_alive_and_valid()
-    if not self.initialized and (all_valid or self.sm.frame * DT_CTRL > 3.0):
+    if not self.initialized and (all_valid or self.sm.frame * DT_CTRL > 3.5):
       self.CI.init(self.CP, self.can_sock, self.pm.sock['sendcan'])
       self.initialized = True
       Params().put_bool("ControlsReady", True)
@@ -477,6 +480,9 @@ class Controls:
       self.LaC.reset()
       self.LoC.reset(v_pid=CS.vEgo)
 
+    if not CS.cruiseState.enabled:
+      self.LoC.reset(v_pid=CS.vEgo)
+
     if not self.joystick_mode:
       # Gas/Brake PID loop
       actuators.gas, actuators.brake, self.v_target, self.a_target = self.LoC.update(self.active, CS, self.CP, long_plan)
@@ -555,8 +561,13 @@ class Controls:
 
     right_lane_visible = self.sm['lateralPlan'].rProb > 0.5
     left_lane_visible = self.sm['lateralPlan'].lProb > 0.5
-    CC.hudControl.rightLaneVisible = bool(right_lane_visible)
-    CC.hudControl.leftLaneVisible = bool(left_lane_visible)
+
+    if self.sm.frame % 100 == 0:
+      self.right_lane_visible = right_lane_visible
+      self.left_lane_visible = left_lane_visible
+
+    CC.hudControl.rightLaneVisible = self.right_lane_visible
+    CC.hudControl.leftLaneVisible = self.left_lane_visible
 
     recent_blinker = (self.sm.frame - self.last_blinker_frame) * DT_CTRL < 5.0  # 5s blinker cooldown
     ldw_allowed = self.is_ldw_enabled and CS.vEgo > LDW_MIN_SPEED and not recent_blinker \
@@ -628,8 +639,6 @@ class Controls:
     controlsState.angleSteers = steer_angle_without_offset * CV.RAD_TO_DEG
     controlsState.cluSpeedMs = self.clu_speed_ms
     controlsState.applyAccel = self.apply_accel
-    controlsState.fusedAccel = self.fused_accel
-    controlsState.leadDist = self.lead_drel
     controlsState.aReqValue = self.aReqValue
     controlsState.aReqValueMin = self.aReqValueMin
     controlsState.aReqValueMax = self.aReqValueMax