+++ selfdrive/controls/lib/lane_planner.py
@@ -1,5 +1,9 @@
 from common.numpy_fast import interp, clip, mean
 import numpy as np
+from cereal import log
+from common.filter_simple import FirstOrderFilter
+from common.numpy_fast import interp
+from common.realtime import DT_MDL
 from selfdrive.hardware import EON, TICI
 from selfdrive.swaglog import cloudlog
 from cereal import log
@@ -26,9 +30,9 @@ class LanePlanner:
     self.ll_x = np.zeros((TRAJECTORY_SIZE,))
     self.lll_y = np.zeros((TRAJECTORY_SIZE,))
     self.rll_y = np.zeros((TRAJECTORY_SIZE,))
-    self.lane_width_estimate = 3.5
-    self.lane_width_certainty = 1.0
-    self.lane_width = 3.5
+    self.lane_width_estimate = FirstOrderFilter(3.7, 9.95, DT_MDL)
+    self.lane_width_certainty = FirstOrderFilter(1.0, 0.95, DT_MDL)
+    self.lane_width = 3.7
 
     self.lll_prob = 0.
     self.rll_prob = 0.
@@ -45,7 +49,6 @@ class LanePlanner:
 
     self.readings = []
     self.frame = 0
-
 
   def parse_model(self, md):
     if len(md.laneLines) == 4 and len(md.laneLines[0].t) == TRAJECTORY_SIZE:
@@ -101,14 +104,15 @@ class LanePlanner:
       # Don't exit dive
       if abs(self.rll_y[0] - self.lll_y[0]) > self.lane_width:
         r_prob = r_prob / interp(l_prob, [0, 1], [1, 3])
+
     else:
       # Find current lanewidth
-      self.lane_width_certainty += 0.05 * (l_prob * r_prob - self.lane_width_certainty)
+      self.lane_width_certainty.update(l_prob * r_prob)
       current_lane_width = abs(self.rll_y[0] - self.lll_y[0])
-      self.lane_width_estimate += 0.005 * (current_lane_width - self.lane_width_estimate)
+      self.lane_width_estimate.update(current_lane_width)
       speed_lane_width = interp(v_ego, [0., 31.], [2.8, 3.5])
-      self.lane_width = self.lane_width_certainty * self.lane_width_estimate + \
-                        (1 - self.lane_width_certainty) * speed_lane_width
+      self.lane_width = self.lane_width_certainty.x * self.lane_width_estimate.x + \
+                        (1 - self.lane_width_certainty.x) * speed_lane_width
 
     clipped_lane_width = min(4.0, self.lane_width)
     path_from_left_lane = self.lll_y + clipped_lane_width / 2.0
@@ -118,7 +122,7 @@ class LanePlanner:
 
     # neokii
     if ENABLE_INC_LANE_PROB and self.d_prob > 0.65:
-      self.d_prob = min(self.d_prob * 1.35, 1.0)
+      self.d_prob = min(self.d_prob * 1.3, 1.0)
 
     lane_path_y = (l_prob * path_from_left_lane + r_prob * path_from_right_lane) / (l_prob + r_prob + 0.0001)
     safe_idxs = np.isfinite(self.ll_t)