+++ selfdrive/ui/qt/offroad/settings.cc
@@ -30,6 +30,8 @@
 #include "selfdrive/ui/qt/widgets/toggle.h"
 #include "selfdrive/ui/ui.h"
 #include "selfdrive/ui/qt/util.h"
+#include "selfdrive/ui/qt/qt_window.h"
+
 
 #include <QComboBox>
 #include <QAbstractItemView>
@@ -107,16 +109,14 @@ TogglesPanel::TogglesPanel(QWidget *parent) : QWidget(parent) {
 DevicePanel::DevicePanel(QWidget* parent) : QWidget(parent) {
   QVBoxLayout *main_layout = new QVBoxLayout(this);
   Params params = Params();
-
-  QString dongle = QString::fromStdString(params.get("DongleId", false));
-  main_layout->addWidget(new LabelControl("Dongle ID", dongle));
+  main_layout->addWidget(new LabelControl("Dongle ID", getDongleId().value_or("N/A")));
   main_layout->addWidget(horizontal_line());
 
   QString serial = QString::fromStdString(params.get("HardwareSerial", false));
   main_layout->addWidget(new LabelControl("Serial", serial));
 
   QHBoxLayout *reset_layout = new QHBoxLayout();
-  reset_layout->setSpacing(30);
+  reset_layout->setSpacing(30); 
 
   // reset calibration button
   QPushButton *reset_calib_btn = new QPushButton("Reset Calibration and LiveParameters");
@@ -132,11 +132,40 @@ DevicePanel::DevicePanel(QWidget* parent) : QWidget(parent) {
     }
   });
 
-  main_layout->addWidget(horizontal_line());
-  main_layout->addLayout(reset_layout);
 
   // Settings and buttons - JPR
-   
+  main_layout->addWidget(horizontal_line());
+  const char* gitpull = "sh /data/openpilot/gitpull.sh";
+  auto gitpullbtn = new ButtonControl("Git Pull and Reboot", "RUN");
+  QObject::connect(gitpullbtn, &ButtonControl::clicked, [=]() {
+    std::system(gitpull);
+    if (ConfirmationDialog::confirm("Process completed successfully. Reboot?", this)){
+      QTimer::singleShot(1000, []() { Hardware::reboot(); });
+    }
+  });
+  main_layout->addWidget(gitpullbtn);
+  main_layout->addWidget(horizontal_line());
+
+  const char* panda_flash = "sh /data/openpilot/panda/board/flash.sh";
+  auto pandaflashbtn = new ButtonControl("Flash Panda Firmware", "RUN");
+  QObject::connect(pandaflashbtn, &ButtonControl::clicked, [=]() {
+    std::system(panda_flash);
+    if (ConfirmationDialog::confirm("Process Completed. Reboot?", this)){
+      QTimer::singleShot(1000, []() { Hardware::reboot(); });
+    }
+  });
+  main_layout->addWidget(pandaflashbtn);
+  main_layout->addWidget(horizontal_line());
+
+  const char* panda_recover = "sh /data/openpilot/panda/board/recover.sh";
+  auto pandarecoverbtn = new ButtonControl("Panda Recover Firmware", "RUN");
+  QObject::connect(pandarecoverbtn, &ButtonControl::clicked, [=]() {
+    std::system(panda_recover);
+    if (ConfirmationDialog::confirm("Process Completed. Reboot?", this)){
+      QTimer::singleShot(1000, []() { Hardware::reboot(); });
+    }
+  });
+  main_layout->addWidget(pandarecoverbtn);
   main_layout->addWidget(horizontal_line());
   auto nTune = new ButtonControl("Run nTune AutoTune for lateral.", "nTune", "Run this after 20 or so miles of driving, to Auto Tune Lateral control.");
   QObject::connect(nTune, &ButtonControl::clicked, [=]() { 
@@ -214,14 +243,16 @@ DevicePanel::DevicePanel(QWidget* parent) : QWidget(parent) {
     });
   }
 
-  auto uninstallBtn = new ButtonControl("Uninstall " + getBrand(), "UNINSTALL");
-  connect(uninstallBtn, &ButtonControl::clicked, [=]() {
-    if (ConfirmationDialog::confirm("Are you sure you want to uninstall?", this)) {
-      Params().putBool("DoUninstall", true);
-    }
-  });
+  ButtonControl *regulatoryBtn = nullptr;
+  if (Hardware::TICI()) {
+    regulatoryBtn = new ButtonControl("Regulatory", "VIEW", "");
+    connect(regulatoryBtn, &ButtonControl::clicked, [=]() {
+      const std::string txt = util::read_file(ASSET_PATH.toStdString() + "/offroad/fcc.html");
+      RichTextDialog::alert(QString::fromStdString(txt), this);
+    });
+  }
 
-  for (auto btn : {dcamBtn, resetCalibBtn, retrainingBtn, uninstallBtn}) {
+  for (auto btn : {dcamBtn, resetCalibBtn, retrainingBtn, regulatoryBtn}) {
     if (btn) {
       main_layout->addWidget(horizontal_line());
       connect(parent, SIGNAL(offroadTransition(bool)), btn, SLOT(setEnabled(bool)));
@@ -234,7 +265,7 @@ DevicePanel::DevicePanel(QWidget* parent) : QWidget(parent) {
   power_layout->setSpacing(30);
 
   QPushButton *reboot_btn = new QPushButton("Reboot");
-  reboot_btn->setStyleSheet("height: 120px;border-radius: 15px; background-color: #393939;");
+  reboot_btn->setObjectName("reboot_btn");
   power_layout->addWidget(reboot_btn);
   QObject::connect(reboot_btn, &QPushButton::clicked, [=]() {
     if (ConfirmationDialog::confirm("Are you sure you want to reboot?", this)) {
@@ -243,7 +274,7 @@ DevicePanel::DevicePanel(QWidget* parent) : QWidget(parent) {
   });
 
   QPushButton *poweroff_btn = new QPushButton("Power Off");
-  poweroff_btn->setStyleSheet("height: 120px;border-radius: 15px; background-color: #E22C2C;");
+  poweroff_btn->setObjectName("poweroff_btn");
   power_layout->addWidget(poweroff_btn);
   QObject::connect(poweroff_btn, &QPushButton::clicked, [=]() {
     if (ConfirmationDialog::confirm("Are you sure you want to power off?", this)) {
@@ -251,6 +282,16 @@ DevicePanel::DevicePanel(QWidget* parent) : QWidget(parent) {
     }
   });
 
+  setStyleSheet(R"(
+    QPushButton {
+      height: 120px;
+      border-radius: 15px;
+    }
+    #reboot_btn { background-color: #393939; }
+    #reboot_btn:pressed { background-color: #4a4a4a; }
+    #poweroff_btn { background-color: #E22C2C; }
+    #poweroff_btn:pressed { background-color: #FF2424; }
+  )");
   main_layout->addLayout(power_layout);
 }
 
@@ -276,10 +317,17 @@ SoftwarePanel::SoftwarePanel(QWidget* parent) : QWidget(parent) {
   QWidget *widgets[] = {versionLbl, lastUpdateLbl, updateBtn, gitBranchLbl, gitCommitLbl, osVersionLbl};
   for (int i = 0; i < std::size(widgets); ++i) {
     main_layout->addWidget(widgets[i]);
-    if (i < std::size(widgets) - 1) {
-      main_layout->addWidget(horizontal_line());
-    }
+    main_layout->addWidget(horizontal_line());
   }
+
+  auto uninstallBtn = new ButtonControl("Uninstall " + getBrand(), "UNINSTALL");
+  connect(uninstallBtn, &ButtonControl::clicked, [=]() {
+    if (ConfirmationDialog::confirm("Are you sure you want to uninstall?", this)) {
+      Params().putBool("DoUninstall", true);
+    }
+  });
+  connect(parent, SIGNAL(offroadTransition(bool)), uninstallBtn, SLOT(setEnabled(bool)));
+  main_layout->addWidget(uninstallBtn);
 
   fs_watch = new QFileSystemWatcher(this);
   QObject::connect(fs_watch, &QFileSystemWatcher::fileChanged, [=](const QString path) {
@@ -405,6 +453,10 @@ QWidget * community_panel() {
 
 //settings - JPR
   toggles_list->addWidget(horizontal_line());
+  toggles_list->addWidget(new ParamControl("PutPrebuilt", "Prebuilt Enable",
+                                  "Create prebuilt files to speed up load time.",
+                                  "../assets/offroad/icon_addon.png"));
+  toggles_list->addWidget(horizontal_line());
   toggles_list->addWidget(new ParamControl("AR",
                                             "Enable Auto Record",
                                             "Starts recording on car start and stops on car off.",
@@ -461,8 +513,12 @@ QWidget * community_panel() {
                                             "warnings: it is beta, be careful!! Openpilot will control the speed of your car",
                                             "../assets/offroad/icon_road.png"
                                               ));
-
- 
+  toggles_list->addWidget(horizontal_line());
+  toggles_list->addWidget(new ParamControl("TPMS_Alerts",
+                                            "Enable TPMS Alerts",
+                                            "Enables Tire Pressure Monitoring System Alerts for Low Tire Pressure.",
+                                            "../assets/offroad/icon_road.png"
+                                              ));
   toggles_list->addWidget(horizontal_line());
   toggles_list->addWidget(new ParamControl("MadModeEnabled",
                                             "Enable HKG MAD mode",
@@ -504,14 +560,6 @@ QWidget * community_panel() {
                                             "Syncs the set speed with the cluster when gas is pressed.",
                                             "../assets/offroad/icon_road.png"
                                             ));
-
-  toggles_list->addWidget(horizontal_line());
-  toggles_list->addWidget(new ParamControl("FuseWithStockScc",
-                                            "Use by fusion with HGK Long",
-                                            "Stock SCC will control gas and OP will control braking, it will smooth out the late abrupt braking from stock SCC.",
-                                            "../assets/offroad/icon_road.png"
-                                            ));
-
   toggles_list->addWidget(horizontal_line());
   toggles_list->addWidget(new ParamControl("ShowDebugUI",
                                             "Show Debug UI",
@@ -572,13 +620,18 @@ SettingsWindow::SettingsWindow(QWidget *parent) : QFrame(parent) {
   // close button
   QPushButton *close_btn = new QPushButton("×");
   close_btn->setStyleSheet(R"(
-    font-size: 140px;
-    padding-bottom: 20px;
-    font-weight: bold;
-    border 1px grey solid;
-    border-radius: 100px;
-    background-color: #292929;
-    font-weight: 400;
+    QPushButton {
+      font-size: 140px;
+      padding-bottom: 20px;
+      font-weight: bold;
+      border 1px grey solid;
+      border-radius: 100px;
+      background-color: #292929;
+      font-weight: 400;
+    }
+    QPushButton:pressed {
+      background-color: #3B3B3B;
+    }
   )");
   close_btn->setFixedSize(200, 200);
   sidebar_layout->addSpacing(45);
@@ -623,6 +676,9 @@ SettingsWindow::SettingsWindow(QWidget *parent) : QFrame(parent) {
       }
       QPushButton:checked {
         color: white;
+      }
+      QPushButton:pressed {
+        color: #ADADAD;
       }
     )").arg(padding));